Splits large module "test_cube_manipulation" (#641)
* Splits large module "test_cube_manipulation" into several smaller modules that will be easier to manage.

* Updates module doc-strings.

* Further doc-string improvements
